file rename: android-service-configuration.ts => adb-wrapper.ts
