+++ src/electron/flux/action/android-setup-actions.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { Action } from 'common/flux/action';
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 
 // This class needs only to represent possible simultaneously invokable actions
 // So, for example, the 'next' action may represent 'continue', 'str again', or 'start scanning'
