+++ src/electron/flux/types/android-setup-state-machine-types.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 import { AndroidSetupStepId } from 'electron/platform/android/setup/android-setup-step-id';
 import { StateMachine } from 'electron/platform/android/setup/state-machine/state-machine';
 import { AndroidSetupActions } from '../action/android-setup-actions';
