+++ src/electron/flux/types/android-setup-store-data.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 import { AndroidSetupStepId } from 'electron/platform/android/setup/android-setup-step-id';
 
 export type AndroidSetupStoreData = {
