+++ src/electron/platform/android/adb-wrapper.ts
