+++ src/electron/platform/android/setup/android-setup-deps.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 
 export type AndroidSetupDeps = {
     hasAdbPath: () => Promise<boolean>;
