+++ src/electron/views/device-connect-view/components/android-setup/device-description.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import { NamedFC } from 'common/react/named-fc';
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 import { css, Icon } from 'office-ui-fabric-react';
 import * as React from 'react';
 import * as styles from './device-description.scss';
