+++ src/electron/views/device-connect-view/components/android-setup/prompt-choose-device-step.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 import {
     CheckboxVisibility,
     DefaultButton,
