+++ src/tests/unit/tests/electron/platform/android/setup/live-android-service-setup-business-logic-factory.test.ts
@@ -5,7 +5,7 @@ import { AndroidServiceApkLocator } from 'electron/platform/android/android-serv
 import {
     AndroidServiceConfigurator,
     AndroidServiceConfiguratorFactory,
-} from 'electron/platform/android/android-service-configurator';
+} from 'electron/platform/android/adb-wrapper';
 import {
     AndroidServiceSetupBusinessLogic,
     LiveAndroidServiceSetupBusinessLogic,
