+++ src/tests/unit/tests/electron/platform/android/setup/steps/prompt-choose-device.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { DeviceInfo } from 'electron/platform/android/android-service-configurator';
+import { DeviceInfo } from 'electron/platform/android/adb-wrapper';
 import { AndroidSetupStepConfigDeps } from 'electron/platform/android/setup/android-setup-steps-configs';
 import { promptChooseDevice } from 'electron/platform/android/setup/steps/prompt-choose-device';
 import { Mock, MockBehavior, Times } from 'typemoq';
